updated phenotypes with phomopsis scores
Phomopsis leaf and cane scores collected by HB and CC June 11, 2020 at Darr added to columns 38, 39, 40, and 41, respectively.  Columns for non-normalized ion concentrations had been removed from the mapping tables previously and that change is also reflected in this update of the phenotypes list so that the correct columns are assigned to everything.